% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_condition.R
\name{assign_condition}
\alias{assign_condition}
\title{Assigns conditions to subjects based on a group assignment matrix.}
\usage{
assign_condition(data, group_labels, assignment_mat, groupvar = "condition")
}
\arguments{
\item{data}{A data frame containing the data to which conditions will be
assigned.}

\item{group_labels}{A vector of treatment group labels. The length of this
vector should be equal to the number of treatment groups in the group
assignment matrix.}

\item{assignment_mat}{A matrix with \code{nsub} rows and \code{ngroups}
columns, where each row represents a subject and each column represents a
treatment group. The matrix specifies which subjects are assigned to each
treatment group.}

\item{groupvar}{A string specifying the name of the variable in \code{data}
that will store the treatment condition assignments. The default value is
\code{"condition"}.}
}
\value{
A data frame with the same columns as \code{data}, with a new column
  added for the treatment condition assignments.
}
\description{
This function assigns treatment conditions to subjects in a dataset based on
a group assignment matrix. The group assignment matrix specifies which
subjects are assigned to each treatment group.
}
