% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_assignment.R
\name{extract_group_assignment}
\alias{extract_group_assignment}
\title{Extracts the group assignments from a genetic algorithm output object.}
\usage{
extract_group_assignment(ga_output, nsub, ngroups)
}
\arguments{
\item{ga_output}{An object of class \code{ga-class}, representing the output
of a genetic algorithm run.}

\item{nsub}{An integer specifying the number of subjects.}

\item{ngroups}{An integer specifying the number of matched groups of N = 1
per condition to which subjects are to be assigned.}
}
\value{
A matrix with \code{nsub} rows and \code{ngroups} columns, where each
  row represents a subject and each column represents a set of matched
  subjects.
}
\description{
Extracts the group assignments from a genetic algorithm output object.
}
