% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_fun_min_f.R
\name{fitness_fun_min_f}
\alias{fitness_fun_min_f}
\title{Calculates the fitness of a genetic algorithm solution for a experimental
design problem with a minimum f-statistic objective.}
\usage{
fitness_fun_min_f(x, nsub, ncond, npg, data)
}
\arguments{
\item{x}{A binary vector representing the genetic algorithm solution.}

\item{nsub}{An integer specifying the number of subjects in the dataset}

\item{ncond}{An integer specifying the number of experimental conditions in
the experiment.}

\item{npg}{An integer specifying the number of subjects per condition.}

\item{data}{A data frame containing the variables on which subjects are to be
matched.}
}
\value{
A numeric value representing the fitness of the genetic algorithm
  solution.
}
\description{
Calculates the fitness of a genetic algorithm solution for a experimental
design problem with a minimum f-statistic objective.
}
