% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_suggestions.R
\name{generate_suggestions}
\alias{generate_suggestions}
\title{Generates suggestions for a genetic algorithm solution.}
\usage{
generate_suggestions(nsuggestions, nsub, ngroups)
}
\arguments{
\item{nsuggestions}{An integer specifying the number of suggestions to
generate.}

\item{nsub}{An integer specifying the number of subjects requiring matching.}

\item{ngroups}{An integer specifying the number of conditions to which
subjects are to be assigned}
}
\value{
A binary matrix with \code{nsuggestions} rows and \code{nsub *
  ngroups} columns, where each row represents a suggested genetic algorithm
  solution.
}
\description{
Generates suggestions for a genetic algorithm solution.
}
\examples{
generate_suggestions(5, 20, 4)

}
