% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_matched_random_assignment.R
\name{perform_matched_random_assignment}
\alias{perform_matched_random_assignment}
\title{Perform matched random assignment}
\usage{
perform_matched_random_assignment(data, ncond, nsuggestions = 500)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the
assignment procedure.}

\item{ncond}{An integer specifying the number of conditions to which subjects
will be assigned.}

\item{nsuggestions}{An optional integer specifying the number of random
suggestions to generate for the GA algorithm. Default is 500.}
}
\value{
A data frame with an additional column indicating the assigned
  condition for each subject.
}
\description{
This function performs matched random assignment of subjects to conditions
using a binary genetic algorithm (GA). The GA algorithm seeks to
pseudo-randomly assign subjects to \code{ncond} groups while minimizing
between group variance on all variables in \code{data}.
}
